enable the ability to export the angular module
https://github.com/marklagendijk/gulp-ng-html2js/issues/15
https://github.com/marklagendijk/gulp-ng-html2js/pull/16

This is an updated PR from the previous issue/pull request based on the
latest master. I don’t expect you to accept it since you already
rejected it, but I’m keeping my own fork of your project with this in
it for my own needs and I’d like to keep the two in sync. Hopefully
others will come along and convince you to apply it. =)  If there is
another way that you’d like this implemented, yet achieve the same
goal, I’m all ears too.